Move getNodeLiveUntilMap to mode_analysis.ts
